snowpipe: rename capped package to pool
