+++ internal/impl/snowflake/output_snowflake_streaming.go
@@ -21,7 +21,7 @@ import (
 	"github.com/redpanda-data/benthos/v4/public/bloblang"
 	"github.com/redpanda-data/benthos/v4/public/service"
 
-	"github.com/redpanda-data/connect/v4/internal/impl/snowflake/capped"
+	"github.com/redpanda-data/connect/v4/internal/impl/snowflake/pool"
 	"github.com/redpanda-data/connect/v4/internal/impl/snowflake/streaming"
 	"github.com/redpanda-data/connect/v4/internal/license"
 )
